+++ src/density.cc
@@ -89,6 +89,11 @@ current_eckart_impl(const ThreeVector &r, const T &plist,
       if (p.get_history().collisions_per_particle == 0) {
         continue;
       }
+      if (par.ignore_unformed()) {
+        if (p.xsec_scaling_factor() < 1) {
+          continue;
+        }
+      }
     }
     const double dens_factor = density_factor(p.type(), dens_type);
     if (std::fabs(dens_factor) < really_small) {