+++ src/tests/listmodus.cc
@@ -594,6 +594,7 @@ TEST(try_create_particle_with_spin_func) {
       1.0,
       false,
       false,
+      false,
       true,
       SpinInteractionType::On,
       std::nullopt};