+++ src/tests/setup.h
@@ -262,6 +262,7 @@ inline ExperimentParameters default_parameters(
       2.5,    // fixed min. cell length
       1.0,    // cross section scaling
       false,  // in thermodynamics outputs spectators are included
+      false,  // ignore unformed particles for thermodynamics
       false,  // do weak decays
       true,   // decay initial particles
       SpinInteractionType::Off,  // no spin interactions